Removal Of Unnecessary Items Removed the duplication of fetches and other components I no longer needed
