--- src/Components/App/useQuery/fetchEpisodes.js
@@ -1,7 +1,0 @@
-const fetchEpisodes = async ({ queryKey }) => {
-  const [, id] = queryKey;
-  const res = await fetch(`${id}`);
-  return res.json();
-};
-
-export default fetchEpisodes;
