--- src/Components/App/useQuery/fetchLocations.js
@@ -1,7 +1,0 @@
-const fetchLocations = async ({ queryKey }) => {
-  const [, id] = queryKey;
-  const res = await fetch(`${id}`);
-  return res.json();
-};
-
-export default fetchLocations;
