+++ src/Components/App/useQuery/fetchSpecificCharacters.js
@@ -1,0 +1,13 @@
+const fetchSpecificUser = async ({ queryKey }) => {
+  const [, id] = queryKey;
+
+  const response = await fetch(
+    `https://rickandmortyapi.com/api/character/${id}`
+  );
+  if (!response.ok) {
+    throw new Error("Network response was not ok");
+  }
+  return response.json();
+};
+
+export default fetchSpecificUser;
