--- src/Components/App/useQuery/fetchSpecificItem.js
@@ -1,6 +1,0 @@
-  async function fetchSpecificItem(urlR) {
-      const res = await fetch({urlR});
-      return res.json();
-  };
-
-  export default fetchSpecificItem;
