+++ src/Components/Styling/App.scss
@@ -14,13 +14,13 @@ body {
 }
 .App {
   margin: 0 auto;
-  
 }
 h1 {
   color: $primary-color;
   font-size: 4em;
   letter-spacing: 2px;
 }
+
 button {
   margin: 0 10px;
   background: transparent;
